pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents[openai,turn-detector,silero,cartesia,deepgram]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2.1",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

# TODO: Remove these once agents 1.2 is released
# If you run into git lfs smudge issues when doing `uv sync`, do this:
# ```
# uv cache clean
# UV_GIT_LFS=1 uv sync
# ```
[tool.uv.sources]
livekit-agents = { git = "https://github.com/livekit/agents.git", branch = "theo/agents1.2", subdirectory = "livekit-agents" }
livekit-plugins-openai = { git = "https://github.com/livekit/agents.git", branch = "theo/agents1.2", subdirectory = "livekit-plugins/livekit-plugins-openai" }
livekit-plugins-turn-detector = { git = "https://github.com/livekit/agents.git", branch = "theo/agents1.2", subdirectory = "livekit-plugins/livekit-plugins-turn-detector" }
livekit-plugins-silero = { git = "https://github.com/livekit/agents.git", branch = "theo/agents1.2", subdirectory = "livekit-plugins/livekit-plugins-silero" }
livekit-plugins-cartesia = { git = "https://github.com/livekit/agents.git", branch = "theo/agents1.2", subdirectory = "livekit-plugins/livekit-plugins-cartesia" }
livekit-plugins-deepgram = { git = "https://github.com/livekit/agents.git", branch = "theo/agents1.2", subdirectory = "livekit-plugins/livekit-plugins-deepgram" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
